+++ src/SimpleClient.php
@@ -1,0 +1,16 @@
+<?php
+namespace MallardDuck\Whois;
+
+/**
+ * A simple concrete version of the Abstract client.
+ *
+ * @author mallardduck <dpock32509@gmail.com>
+ *
+ * @copyright lucidinternets.com 2018
+ *
+ * @version 0.6.0
+ */
+class SimpleClient extends AbstractWhoisClient
+{
+
+}