+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New basic version of client as SimpleClient. Just a concrete version of the AbstractWhoisClient.
+
 ### Changed
 - Fixed issues with PHPunit XML config.
 - Modify $lastMatch to be the exact match, rather than an array containing it.